+++ starters/shop/next.config.ts
@@ -2,7 +2,7 @@ import type { NextConfig } from "next";
 import { initOpenNextCloudflareForDev } from "@opennextjs/cloudflare";
 
 const nextConfig: NextConfig = {
-  transpilePackages: ["@kumooo/ui", "@kumooo/framework"],
+  transpilePackages: ["@kumooo/ui", "@kumooo/theme-packs"],
 };
 
 export default nextConfig;